requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
SQLAlchemy==2.0.31
Werkzeug==3.1.3
python-dotenv==1.0.1
Pillow==11.1.0
requests==2.32.3
qrcode[pil]==8.0
gunicorn==23.0.0
